tests/wpt/tests/html/semantics/permission-element/geolocation-element/parent-image-mask.tentative.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Geolocation Element: invalid if parent has mask-image</title>
    <link
      rel="help"
      href="https://github.com/WICG/PEPC/blob/main/explainer.md#locking-the-pepc-style"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      async_test((t) => {
        const parent = document.createElement("div");
        const element = document.createElement("geolocation");
        element.setAttribute("type", "camera");
        parent.appendChild(element);

        parent.style.maskImage = "url(mask.png)";

        element.onvalidationstatuschange = t.step_func(() => {
          // These two invalid reasons are expected when the geolocation element was just created.
          if (
            element.invalidReason == "unsuccessful_registration" ||
            element.invalidReason == "intersection_changed"
          ) {
            return;
          }
          assert_equals(
            element.invalidReason,
            "style_invalid",
            `Parent with mask-image should be invalid`,
          );
          t.done();
        });

        document.body.appendChild(parent);
      }, "Geolocation element is invalid if it has a parent with mask-image.");
    </script>
  </body>
</html>
